app/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
bokeh==3.2.2
cachetools==5.3.2
Cartopy==0.22.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
citipy==0.0.6
click==8.1.7
colorama==0.4.6
colorcet==3.0.1
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dnspython==2.4.2
et-xmlfile==1.1.0
executing==2.0.0
fastjsonschema==2.18.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
gast==0.5.4
geoviews==1.10.1
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
holoviews==1.17.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hvplot==0.8.4
idna==3.4
imageio==2.33.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kdtree==0.16
keras==2.15.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
linkify-it-py==2.0.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
openai==1.3.8
opencv-python==4.9.0.80
openpyxl==3.1.0
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
panel==1.2.3
param==1.13.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pymongo==4.6.0
pyparsing==3.0.9
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pyviz_comms==3.0.0
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
selenium==4.14.0
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
splinter==0.19.0
SQLAlchemy==2.0.22
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2024.1.30
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
trio==0.22.2
trio-websocket==0.11.1
ttkbootstrap==1.10.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uc-micro-py==1.0.2
undetected-chromedriver==3.5.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xyzservices==2023.10.0
